LA-227: Add PublicInformation object

+++ src/main/java/com/selina/lending/internal/dto/creditcommitments/response/DetailResponseDto.java
@@ -54,6 +54,6 @@ public class DetailResponseDto {
     String type;
     Double currentBalance;
     String interestOnlyBalance;
-    Integer amount;
+    Double amount;
     String code;
 }
